requirements.txt
libsass
django-npm
